+++ hive/test/sentry_hive_impl_test.dart
@@ -9,6 +9,7 @@ import 'package:sentry/src/sentry_tracer.dart';
 import 'package:sentry_hive/src/sentry_box.dart';
 import 'package:sentry_hive/src/sentry_hive_impl.dart';
 import 'package:sentry_hive/src/sentry_lazy_box.dart';
+import 'package:sentry_hive/src/version.dart';
 import 'package:test/test.dart';
 
 import 'mocks/mocks.mocks.dart';
@@ -45,10 +46,11 @@ void main() {
 
     setUp(() async {
       fixture = Fixture();
-      await fixture.setUp();
 
       when(fixture.hub.options).thenReturn(fixture.options);
       when(fixture.hub.getSpan()).thenReturn(fixture.tracer);
+
+      await fixture.setUp();
     });
 
     tearDown(() async {
@@ -113,11 +115,12 @@ void main() {
 
     setUp(() async {
       fixture = Fixture();
-      await fixture.setUp(injectMockHive: true);
 
       when(fixture.hub.options).thenReturn(fixture.options);
       when(fixture.hub.getSpan()).thenReturn(fixture.tracer);
       when(fixture.mockHive.close()).thenAnswer((_) async => {});
+
+      await fixture.setUp(injectMockHive: true);
     });
 
     test('throwing boxExists adds error span', () async {
@@ -267,6 +270,40 @@ void main() {
         'openLazyBox',
         fixture.getCreatedSpan(),
         fixture.exception,
+      );
+    });
+  });
+
+  group('integrations', () {
+    late Fixture fixture;
+
+    setUp(() async {
+      fixture = Fixture();
+
+      when(fixture.hub.options).thenReturn(fixture.options);
+      when(fixture.hub.getSpan()).thenReturn(fixture.tracer);
+
+      await fixture.setUp();
+    });
+
+    tearDown(() async {
+      await fixture.tearDown();
+    });
+
+    test('adds integration', () {
+      expect(
+        fixture.options.sdk.integrations.contains('SentryHiveTracing'),
+        true,
+      );
+    });
+
+    test('adds package', () {
+      expect(
+        fixture.options.sdk.packages.any(
+          (element) =>
+              element.name == packageName && element.version == sdkVersion,
+        ),
+        true,
       );
     });
   });